requirements.txt
pandas==2.0.3
psycopg2-binary==2.9.9
boto3==1.28.85
python-dotenv==1.0.0
Faker==19.13.0
